git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2778 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/resource.c
@@ -971,7 +971,7 @@ MagickExport MagickBooleanType ResourceComponentGenesis(void)
 #endif
 #if defined(MAGICKCORE_HAVE_GETDTABLESIZE) && defined(MAGICKCORE_POSIX_SUPPORT)
   if (files < 0)
-    files=getdtablesize();
+    files=(ssize_t) getdtablesize();
 #endif
   if (files < 0)
     files=64;
